Refine Class Pass and Schedule Page UI with Responsive Enhancements
- Update class pass content with improved responsive typography and spacing
- Enhance mobile and desktop layout with adaptive font sizes and padding
- Optimize class card and schedule page components with dynamic styling
- Improve visual consistency across different screen sizes
- Add subtle shadow and border effects for better depth and readability
- Adjust text and icon sizes for improved legibility on various devices
