+++ src/app/class-pass/class-pass-content.tsx
@@ -123,7 +123,7 @@ function CountdownTimer({
 
   if (isInProgress) {
     return (
-      <span className="text-green-600 font-semibold text-lg">
+      <span className="text-gray-700 text-[10px] sm:text-base md:text-lg">
         {language === "en" ? "In progress" : "En progreso"}
       </span>
     );
@@ -132,7 +132,7 @@ function CountdownTimer({
   if (!timeLeft) return null;
 
   return (
-    <span className="text-gray-700 text-lg">
+    <span className="text-gray-700 text-[10px] sm:text-base md:text-lg">
       {language === "en" ? (
         <>
           Starts in {timeLeft.hours > 0 ? `${timeLeft.hours}h ` : ""}
@@ -208,7 +208,7 @@ function ClassPassSkeletonLoader() {
             </motion.div>
 
             {/* Ver Horarios button skeleton */}
-            <Skeleton className="h-14 w-full rounded-xl" />
+            <Skeleton className="h-14 w-full rounded-xl shadow-lg" />
           </div>
         </div>
       </div>
@@ -557,7 +557,7 @@ export function ClassPassContent() {
       transition={{ duration: 0.5 }}
       className="min-h-screen bg-gradient-to-b from-blue-50 to-white pt-14"
     >
-      <div className="container mx-auto px-2 sm:px-4 py-2 sm:py-16 md:py-24">
+      <div className="container mx-auto px-2 sm:px-4 md:px-4 py-2 sm:py-16 md:py-24">
         <div className="max-w-4xl mx-auto">
           <div className="space-y-2 sm:space-y-4">
             {nextClass && singleClassPass && (
@@ -566,12 +566,12 @@ export function ClassPassContent() {
                   initial={{ y: 20, opacity: 0 }}
                   animate={{ y: 0, opacity: 1 }}
                   transition={{ duration: 0.5, delay: 0.2 }}
-                  className="bg-white rounded-xl sm:rounded-2xl shadow-lg overflow-hidden border border-gray-100 mb-2 sm:mb-4"
+                  className="bg-white rounded-xl sm:rounded-2xl md:rounded-2xl shadow-lg overflow-hidden border border-gray-100 mb-2 sm:mb-4"
                 >
                   <div className="p-2 sm:p-6 md:p-8">
                     {/* Header Section with Class Info */}
-                    <div className="flex flex-col gap-1.5 sm:gap-4 pb-2 sm:pb-6 border-b border-gray-100">
-                      <div className="flex flex-row items-start justify-between gap-2 sm:gap-6">
+                    <div className="flex flex-col gap-1.5 sm:gap-4 md:gap-4 pb-2 sm:pb-6 md:pb-6 border-b border-gray-100">
+                      <div className="flex flex-row items-start justify-between gap-2 sm:gap-6 md:gap-6">
                         {/* Left side - Class info */}
                         <div className="flex flex-col flex-1 min-w-0">
                           <motion.h2
@@ -586,7 +586,7 @@ export function ClassPassContent() {
                             initial={{ x: -20, opacity: 0 }}
                             animate={{ x: 0, opacity: 1 }}
                             transition={{ duration: 0.5, delay: 0.4 }}
-                            className="text-xs sm:text-xl md:text-2xl text-gray-600 font-medium mt-0.5 sm:mt-2"
+                            className="text-xs sm:text-xl md:text-2xl text-gray-600 font-medium mt-0.5 sm:mt-2 md:mt-2"
                           >
                             {language === "en" ? "with" : "con"}{" "}
                             <span className="font-semibold">
@@ -597,7 +597,7 @@ export function ClassPassContent() {
                             initial={{ x: -20, opacity: 0 }}
                             animate={{ x: 0, opacity: 1 }}
                             transition={{ duration: 0.5, delay: 0.5 }}
-                            className="flex flex-wrap items-center gap-1 sm:gap-3 text-[10px] sm:text-lg md:text-xl mt-1 sm:mt-3"
+                            className="flex flex-wrap items-center gap-1 sm:gap-3 md:gap-3 text-[10px] sm:text-lg md:text-xl mt-1 sm:mt-3 md:mt-3"
                           >
                             <p className="font-medium text-gray-800">
                               {`${nextClass.day} ${nextClass.time}`}
@@ -614,14 +614,14 @@ export function ClassPassContent() {
                           initial={{ x: 20, opacity: 0 }}
                           animate={{ x: 0, opacity: 1 }}
                           transition={{ duration: 0.5, delay: 0.3 }}
-                          className="flex flex-col items-end justify-start gap-1 sm:gap-2"
+                          className="flex flex-col items-end justify-start gap-1 sm:gap-2 md:gap-2"
                         >
                           <div className="text-base sm:text-3xl md:text-4xl font-bold bg-gradient-to-r from-green-600 to-teal-600 bg-clip-text text-transparent tracking-tight">
                             S/. {Number(singleClassPass.price).toFixed(2)}
                           </div>
-                          <div className="bg-gradient-to-r from-green-600/10 to-teal-600/10 px-1.5 sm:px-4 py-0.5 sm:py-2 rounded-lg flex items-center gap-1 sm:gap-2">
-                            <Tag className="w-3 h-3 sm:w-4 sm:h-4 text-green-600" />
-                            <span className="text-green-700 font-medium text-[10px] sm:text-base whitespace-nowrap">
+                          <div className="bg-gradient-to-r from-green-600/10 to-teal-600/10 px-1.5 sm:px-4 md:px-4 py-0.5 sm:py-2 md:py-2 rounded-lg flex items-center gap-1 sm:gap-2 md:gap-2">
+                            <Tag className="w-3 h-3 sm:w-4 sm:h-4 md:w-4 md:h-4 text-green-600" />
+                            <span className="text-green-700 font-medium text-[10px] sm:text-base md:text-base whitespace-nowrap">
                               {nextClass?.activity
                                 .toLowerCase()
                                 .includes("acro")
@@ -640,12 +640,12 @@ export function ClassPassContent() {
                       initial={{ opacity: 0, y: 20 }}
                       animate={{ opacity: 1, y: 0 }}
                       transition={{ duration: 0.5, delay: 0.6 }}
-                      className="mt-2 sm:mt-6"
+                      className="mt-2 sm:mt-6 md:mt-6"
                     >
                       <button
                         onClick={handleClassSelection}
                         disabled={isBooking}
-                        className="relative w-full bg-gradient-to-r from-green-600 to-teal-600 text-white py-2 sm:py-6 px-2 sm:px-6 rounded-lg sm:rounded-xl transition-all duration-300 flex items-center justify-center gap-1.5 sm:gap-3 text-sm sm:text-xl md:text-2xl font-semibold shadow-lg hover:shadow-xl group overflow-hidden"
+                        className="relative w-full bg-gradient-to-r from-green-600 to-teal-600 text-white py-2 sm:py-6 md:py-6 px-2 sm:px-6 md:px-6 rounded-lg sm:rounded-xl md:rounded-xl transition-all duration-300 flex items-center justify-center gap-1.5 sm:gap-3 md:gap-3 text-sm sm:text-xl md:text-2xl font-semibold shadow-lg hover:shadow-xl group overflow-hidden"
                       >
                         {/* Darkening overlay */}
                         <motion.div
@@ -701,9 +701,9 @@ export function ClassPassContent() {
                     </motion.div>
 
                     {/* Additional Info */}
-                    <div className="mt-2 pt-2 sm:mt-6 sm:pt-6 border-t border-gray-100">
-                      <div className="flex flex-col sm:grid sm:grid-cols-3 gap-1 sm:gap-0">
-                        <div className="flex items-center gap-1.5 sm:gap-3 justify-start">
+                    <div className="mt-2 pt-2 sm:mt-6 md:mt-6 sm:pt-6 md:pt-6 border-t border-gray-100">
+                      <div className="grid grid-cols-3 items-center">
+                        <div className="flex items-center gap-1.5 sm:gap-3 md:gap-3 justify-start">
                           <Users className="w-3.5 h-3.5 sm:w-5 sm:h-5 md:w-6 md:h-6 text-green-500" />
                           <span className="text-gray-700 text-[10px] sm:text-base md:text-lg">
                             {(nextClass.timeSlot?.sessionType?.maxConsumers || 12) - nextClass.currentReservations}/
@@ -716,13 +716,13 @@ export function ClassPassContent() {
                                 exit={{ y: -10, opacity: 0 }}
                                 transition={{ duration: 0.2 }}
                               >
-                                {language === "en" ? "mats" : "mats"}
+                                {language === "en" ? "mats available" : "mats disponibles"}
                               </motion.span>
                             </AnimatePresence>
                           </span>
                         </div>
 
-                        <div className="flex items-center gap-1.5 sm:gap-3 justify-start sm:justify-center">
+                        <div className="flex items-center gap-1.5 sm:gap-3 md:gap-3 justify-start sm:justify-center">
                           <Clock className="w-3.5 h-3.5 sm:w-5 sm:h-5 md:w-6 md:h-6 text-green-500" />
                           <CountdownTimer
                             classTime={nextClass.time}
@@ -730,10 +730,10 @@ export function ClassPassContent() {
                           />
                         </div>
 
-                        <div className="flex items-center gap-1.5 sm:gap-3 justify-start sm:justify-end">
+                        <div className="flex items-center gap-1.5 sm:gap-3 md:gap-3 justify-start sm:justify-end">
                           <Globe2 className="w-3.5 h-3.5 sm:w-5 sm:h-5 md:w-6 md:h-6 text-green-500" />
                           <span className="text-gray-700 text-[10px] sm:text-base md:text-lg whitespace-nowrap">
-                            {language === "en" ? "Bilingual" : "Bilingüe"}
+                            {language === "en" ? "Bilingual Class" : "Clase Bilingüe"}
                           </span>
                         </div>
                       </div>
@@ -745,7 +745,11 @@ export function ClassPassContent() {
                 <motion.button
                   onClick={handleViewSchedule}
                   disabled={isViewingSchedule}
-                  className={`w-full bg-white/90 backdrop-blur-sm text-gray-500 py-2 sm:py-4 px-2.5 sm:px-6 rounded-lg sm:rounded-xl hover:bg-white/95 transition-all duration-300 flex items-center justify-center gap-1.5 sm:gap-2 border border-gray-200 font-semibold text-xs sm:text-lg shadow-sm hover:shadow-md group ${isViewingSchedule ? 'cursor-not-allowed opacity-80' : ''}`}
+                  className={`w-full bg-white/90 backdrop-blur-sm text-gray-500 py-2 sm:py-4 md:py-4 px-2.5 sm:px-6 md:px-6 rounded-lg sm:rounded-xl md:rounded-xl hover:bg-white/95 transition-all duration-300 flex items-center justify-center gap-1.5 sm:gap-2 md:gap-2 border border-gray-200 font-semibold text-xs sm:text-lg md:text-lg group ${isViewingSchedule ? 'cursor-not-allowed opacity-80' : ''}`}
+                  style={{
+                    boxShadow: "0 4px 6px -1px rgba(0, 0, 0, 0.05), 0 2px 4px -1px rgba(0, 0, 0, 0.03)",
+                    borderBottom: "4px solid #f3f4f6"
+                  }}
                 >
                   <AnimatePresence mode="wait">
                     {isViewingSchedule ? (
@@ -754,9 +758,9 @@ export function ClassPassContent() {
                         initial={{ opacity: 0 }}
                         animate={{ opacity: 1 }}
                         exit={{ opacity: 0 }}
-                        className="flex items-center gap-1.5 sm:gap-2"
+                        className="flex items-center gap-1.5 sm:gap-2 md:gap-2"
                       >
-                        <Loader2 className="w-3 h-3 sm:w-5 sm:h-5 animate-spin" />
+                        <Loader2 className="w-3 h-3 sm:w-5 sm:h-5 md:w-5 md:h-5 animate-spin" />
                         <span>
                           {language === "en" ? "Loading..." : "Cargando..."}
                         </span>
@@ -774,7 +778,7 @@ export function ClassPassContent() {
                     )}
                   </AnimatePresence>
                   {!isViewingSchedule && (
-                    <ChevronRight className="w-3 h-3 sm:w-5 sm:h-5 transition-transform group-hover:translate-x-1" />
+                    <ChevronRight className="w-3 h-3 sm:w-5 sm:h-5 md:w-5 md:h-5 transition-transform group-hover:translate-x-1" />
                   )}
                 </motion.button>
               </div>
